+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Added support for the Write API in the mock server.
+
+
 ## [1.14.1] - 2024-01-14
 ### Fixed
 * Update `path-to-regexp` dependency due to CVE-2024-52798
@@ -196,6 +201,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-mock/compare/v1.14.1...HEAD
 [1.14.1]: https://github.com/DeepLcom/deepl-mock/compare/v1.14.0...v1.14.1
 [1.14.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.12.0...v1.13.0